requirements.txt
awesome-slugify==1.6.5
django==1.10.1
echarts-python==0.1.3
future==0.16.0
Jinja2==2.10
jupyter-echarts-pypkg==0.1.0
lml==0.0.2
MarkupSafe==1.0
numpy==1.14.4
paho-mqtt==1.3.1
Pillow==5.1.0
pyecharts==0.4.1
pyecharts-jupyter-installer==0.0.3
pymongo==3.6.0
pymssql==1.0.3
pymysql==0.8.0
pyodbc==4.0.22
pytz==2017.3
regex==2018.6.6
style==1.1.0
Unidecode==0.4.21
UNKNOWN==0.0.0
update==0.0.1
virtualenv==16.0.0
xlwt==1.3.0
